Request generic /?art from server
Previously we would just create a bunch of fallback background images on
the client side. This results in many 404s which is undesirable.

The way we do it now is to request .../?art of the same directory of the
song. This retuns a permanent redirect to a file the server has found.

It also doesn't unload the old art if the next one hasn't loaded yet
(except if you skip ahead really fast).